+++ src/ServerPrivate.hh
@@ -31,6 +31,7 @@
 #include <sdf/Root.hh>
 
 #include <ignition/common/SignalHandler.hh>
+#include <ignition/common/URI.hh>
 #include <ignition/common/WorkerPool.hh>
 
 #include <ignition/fuel_tools/FuelClient.hh>
@@ -87,6 +88,11 @@ namespace ignition
       /// \param[in] _uri The resource URI to fetch.
       /// \return Path to the downloaded resource, empty on error.
       public: std::string FetchResource(const std::string &_uri);
+
+      /// \brief Fetch a resource from Fuel using fuel-tools.
+      /// \param[in] _uri The resource URI to fetch.
+      /// \return Path to the downloaded resource, empty on error.
+      public: std::string FetchResourceUri(const common::URI &_uri);
 
       /// \brief Signal handler callback
       /// \param[in] _sig The signal number