+++ Changelog.md
@@ -4,8 +4,11 @@
 
 ### Ignition Gazebo 4.0.0 (20XX-XX-XX)
 
-1. Support <actor><pose> and <actor><plugin>
-    * [Pull Request 542](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/542)
+1. Add support for visual visibility flags and camera visibility mask
+    * [pull request 559](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/559)
+
+1. support <actor><pose> and <actor><plugin>
+    * [pull request 542](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/542)
 
 1. Depend on ign-rendering4, ign-gui4, ign-sensors4
     * [Pull Request 540](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/540)