+++ Changelog.md
@@ -1,9 +1,12 @@
 ## Ignition Gazebo 2.x
 
-### Ignition Gazebo 2.x.x
+### Ignition Gazebo 2.xx.xx (20XX-XX-XX)
 
 1.  Add DetachableJoint: A system that initially attaches two models via a fixed joint and allows for the models to get detached during simulation via a topic.
     * [Pull Request 440](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/440)
+
+1. Add `JointPositionReset` and `JointVelocityReset` components to reset the joint state.
+    * [Pull Request 437](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/437)
 
 ### Ignition Gazebo 2.15.0 (2020-02-07)