+++ Changelog.md
@@ -1,9 +1,12 @@
 ## Ignition Gazebo 2.x
 
-### Ignition Gazebo 2.xx.xx (20xx-xx-xx)
+### Ignition Gazebo 2.xx.xx (20XX-XX-XX)
 
 1. Add support for setting visual transparency through SDF
     * [Pull Request 547](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/547)
+
+1. Add `JointPositionReset` and `JointVelocityReset` components to reset the joint state.
+    * [Pull Request 437](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/437)
 
 ### Ignition Gazebo 2.15.0 (2020-02-07)